building module desired outputs
when testing if the outputs of the building module are correct, use the 'address_surr' and 'surr_out' signals.
The last value surr_out has before address_surr changes value should correspond with the right table in the file.
The table is correct if the seed is 522. (same conditions as in building_main2_tb behaviour).